fix: resolve 9 bugs in Manage Bids page and live bidding system
Bug B1: Fixed hardcoded 'Avg Time Between Bids' — now calculated from real bid timestamps
Bug B2: Fixed 'Avg Bid Amount' — now averages ALL individual bids, not just highest per auction
Bug B3: Added working search functionality to admin bids page (filter by bidder, auction, amount)
Bug B4: Added pagination to bids list (20 per page with prev/next navigation)
Bug B5: Fixed wallet concurrency bug — added pessimistic lock on user row (findByIdForUpdate)
Bug B6: Fixed bid form showing on expired auctions — added isExpired() check in template
Bug B7: Fixed bid history sort order — now sorted by timestamp DESC (newest first)
Bug B8: Fixed BidDTO validation — changed double to BigDecimal for @DecimalMin compatibility
Bug B9: Added notification popover to 6 admin pages (bids, users, auctions, analytics, settings, dashboard)

Also: Updated 'Highest' badge in bid history to check actual highest bid, not list position

+++ src/main/java/com/auctionhouse/controller/BidController.java
@@ -55,7 +55,7 @@ public class BidController {
             User bidder = userService.getCurrentUser(userDetails.getUsername());
 
             // Place the bid (additional validation happens in service layer)
-            Bid bid = bidService.placeBid(bidDTO.getAuctionId(), bidder, bidDTO.getAmount());
+            Bid bid = bidService.placeBid(bidDTO.getAuctionId(), bidder, bidDTO.getAmount().doubleValue());
 
             redirectAttributes.addFlashAttribute("success",
                     String.format("Bid of $%,.2f placed successfully!", bid.getAmount()));
